moved types where the should be
