+++ hendricks-bot/src/types/Plugin.d.ts
