+++ hendricks-bot/src/types/PluginLoader.d.ts
@@ -1,4 +1,4 @@
-import Plugin from '../Plugin';
+import Plugin from './Plugin';
 
 export default interface PluginLoader {
 	loadPlugins(): Promise<Plugin[]>;
